pyproject.toml
[tool.poetry]
name = "repl-nix-chatbuddyai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
openai = "^1.44.0"
psycopg2-binary = "^2.9.9"
flask-session = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
